rename reset command
